+++ Mario_Game/modles/SaveData.cs
@@ -11,6 +11,6 @@ namespace Mario_Game.modles
     {
         public Vector2 SavedPostion { get; set; }
         public Vector2 SavedAnimation { get; set; }
-        public int SavedCoins { get; set; }
+        public float SavedCoins { get; set; }
     }
 }
